+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 0.9.4
-
+- Add yes/no dialog for logout
 
 ## 0.9.3
 - Add settings screen: