+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.9.4
 - Add yes/no dialog for logout
+- Add settings for the password strength
 
 ## 0.9.3
 - Add settings screen: